requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dacite==1.9.2
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_colorscales==0.0.4
dash_daq==0.5.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dtale==3.16.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
Flask-Compress==1.17
fonttools==4.55.8
fqdn==1.5.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.8
lz4==4.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.1.1
multimethod==1.12
narwhals==1.25.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.1.3
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.8.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
squarify==0.4.4
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.42.0
strsimpy==0.2.1
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
visions==0.7.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wordcloud==1.9.4
xarray==2025.1.2
xlrd==2.0.1
ydata-profiling==4.12.2
zipp==3.21.0
zstandard==0.23.0
